More data. refactored data_imports. Integer -> SmallIntger, mostly

+++ icw/api/db/casualty.py
@@ -1,5 +1,4 @@
-from geoalchemy2 import Geometry
-from sqlalchemy import String, Column, Integer, ForeignKey, ForeignKeyConstraint
+from sqlalchemy import String, Column, ForeignKey, ForeignKeyConstraint, SmallInteger
 from sqlalchemy.orm import relationship
 
 from . import Base
@@ -9,18 +8,18 @@ class Casualty(Base):
     __tablename__ = 'casualty'
 
     accident_id = Column(String(13), ForeignKey('accident.id'), index=True, primary_key=True)
-    vehicle_ref = Column(Integer, primary_key=True)
-    casualty_ref = Column(Integer, primary_key=True)
+    vehicle_ref = Column(SmallInteger, primary_key=True)
+    casualty_ref = Column(SmallInteger, primary_key=True)
 
-    class_id = Column(Integer, ForeignKey('casualty_class.id'), index=True)
-    sex_id = Column(Integer, ForeignKey('sex.id'), index=True)
-    severity_id = Column(Integer, ForeignKey('casualty_severity.id'), index=True)
-    age_band_id = Column(Integer, ForeignKey('age_band.id'), index=True)
-    type_id = Column(Integer, ForeignKey('vehicle_type.id'), index=True)
-    age = Column(Integer, nullable=True, index=True)
+    class_id = Column(SmallInteger, ForeignKey('casualty_class.id'), index=True)
+    sex_id = Column(SmallInteger, ForeignKey('sex.id'), index=True)
+    severity_id = Column(SmallInteger, ForeignKey('casualty_severity.id'), index=True)
+    age_band_id = Column(SmallInteger, ForeignKey('age_band.id'), index=True)
+    type_id = Column(SmallInteger, ForeignKey('vehicle_type.id'), index=True)
+    age = Column(SmallInteger, nullable=True, index=True)
 
-    pedestrian_location_id = Column(Integer, ForeignKey('pedestrian_location.id'), nullable=True)
-    pedestrian_movement_id = Column(Integer, ForeignKey('pedestrian_movement.id'), nullable=True)
+    pedestrian_location_id = Column(SmallInteger, ForeignKey('pedestrian_location.id'), nullable=True)
+    pedestrian_movement_id = Column(SmallInteger, ForeignKey('pedestrian_movement.id'), nullable=True)
 
     accident = relationship('Accident', viewonly=True)
     vehicle = relationship('Vehicle', foreign_keys=[accident_id, vehicle_ref], backref='casualties')
